// frontend/src/api/models/cadidates.ts
export interface Candidate {
    id: number;
    name: string;
    phone: string;
    email: string;
    contacts: string;
    skills: string[];
    experience: number;
    position: string;
    grade: string;
    speciality: string;
    education: string;
    summary: string;
    resumeLink: string;
    isCold: boolean;
}